working on request model for favourites- corresponds to plans to have favourites as the user types in the input
