+++ src/main/java/Entities/Favourites.java
@@ -2,13 +2,13 @@ package Entities;
 import java.util.ArrayList;
 
 public class Favourites{
-    private ArrayList<Song> favourites;
+    private ArrayList<String> favourites;
 
-    public Favourites(ArrayList<Song> favourites){
+    public Favourites(ArrayList<String> favourites){
         this.favourites = favourites;
     }
 
-    public ArrayList<Song> getFavourites(){
+    public ArrayList<String> getFavourites(){
         return this.favourites;
     }
