+++ src/main/java/Presenters/HistoryPresenter.java
@@ -1,11 +1,14 @@
 package Presenters;
 
 import Processors.HistoryProcessor;
+import UI.HistoryWindow;
 
 import java.util.ArrayList;
 
 public class HistoryPresenter {
     public HistoryPresenter(){
-        ArrayList<String> tabledata = HistoryProcessor.get_song_strings(HistoryProcessor.userAccount);
+        String [][] tabledata = HistoryProcessor.toArray(HistoryProcessor.userAccount);
+        HistoryWindow.Hwindow(tabledata);
     }
+
 }
